Semana08Sesion02/practiceORM/backend/requirements.txt
backpack==0.1
blinker==1.6.2
cleo==0.6.8
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
Faker==0.8.18
Flask==2.3.1
Flask-Cors==3.0.10
flask-orator==0.2.0
inflection==0.3.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
orator==0.9.9
pastel==0.1.1
pendulum==1.5.1
psycopg2==2.9.6
pyaml==16.12.2
Pygments==2.15.1
pylev==1.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytzdata==2020.1
PyYAML==5.4.1
simplejson==3.19.1
six==1.16.0
text-unidecode==1.2
tzlocal==1.5.1
Werkzeug==2.3.2
wrapt==1.15.0
